requirements.txt
langchain>=0.1.0
langchain-community>=0.0.30
langchain-anthropic>=0.1.0
langchain-aws>=0.1.0
boto3>=1.34.0
openai>=1.0.0
anthropic>=0.18.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0  # Environment variable management

# AI Tools Dependencies
six>=1.16.0  # For Python 2/3 compatibility modernization
pyupgrade>=3.10.0  # Reference implementation for upgrade patterns
modernize>=0.8.0  # Reference implementation for 2/3 compatibility
autopep8>=2.0.0  # Code formatting capabilities
black>=23.0.0  # Modern code formatting
libcst>=1.0.0  # Code syntax tree manipulation
